+++ src/core/core.cpp
@@ -736,7 +736,7 @@ void print_diagnosis(LONG Signal)
          }
          else class_name = "None";
 
-         fprintf(fd, "  Object Context: #%d / %p [Class: %s / $%.8x]\n", ctx->object()->UID, ctx->object(), class_name, class_id);
+         fprintf(fd, "  Object Context: #%d / %p [Class: %s / $%.8x]\n", ctx->object()->UID, ctx->object(), class_name, ULONG(class_id));
       }
 
       glPageFault = 0;
